requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
charset-normalizer==2.1.1
colorama==0.4.4
colorlog==6.6.0
debugpy==1.6.0
decorator==5.1.1
discord-components-mirror==2.1.5
discord.py==2.0.1
entrypoints==0.4
executing==0.8.3
frozenlist==1.3.1
hikari==2.0.0.dev110
idna==3.3
ipykernel==6.13.0
ipython==8.3.0
jedi==0.18.1
jupyter-client==7.3.1
jupyter-core==4.10.0
matplotlib-inline==0.1.3
multidict==6.0.2
nest-asyncio==1.5.5
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.29
psutil==5.9.0
pure-eval==0.2.2
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pywin32==304
pyzmq==22.3.0
six==1.16.0
stack-data==0.2.0
tornado==6.1
traitlets==5.2.0
wcwidth==0.2.5
yarl==1.8.1
